Added autorate, debugged neural network, issue : need to debug rotate

+++ Imagery/Rotations/rotations.h
@@ -8,7 +8,7 @@
 
 void rotate(Image *image, double angle);
 
-float detectDiffAngle(Image *image, float precision);
+double detectDiffAngle(Image *image, float precision);
 
 void autoRotate(Image *image, float precision);
